+++ src/parser/enrichers/DDBFeatureEnricher.js
@@ -116,6 +116,7 @@ import MagicalCunning from "./feature/warlock/MagicalCunning.js";
 import FiendishResilience from "./feature/warlock/FiendishResilience.js";
 import ClairvoyantCombatant from "./feature/warlock/ClairvoyantCombatant.js";
 import CreateThrall from "./feature/warlock/CreateThrall.js";
+import ArcaneWard from "./feature/wizard/ArcaneWard.js";
 
 export default class DDBFeatureEnricher extends DDBBaseEnricher {
   constructor() {
@@ -247,6 +248,7 @@ export default class DDBFeatureEnricher extends DDBBaseEnricher {
     "Fiendish Resilience": () => FiendishResilience,
     "Clairvoyant Combatant": () => ClairvoyantCombatant,
     "Create Thrall": () => CreateThrall,
+    "Arcane Ward": () => ArcaneWard,
   };
 
   NAME_HINTS_2014 = {